@@ -8,8 +8,9 @@
     "fastify": "^5.1.0",
     "lodash": "^4.17.21",
     "platform": "^1.3.6",
-    "react": "^18.3.1",
-    "tgui": "workspace:*"
+    "react": "^19.1.0",
+    "tgui": "workspace:*",
+    "tgui-core": "^3.1.4"
   },
   "devDependencies": {
     "@types/react": "^19.1.0",
