@@ -10,8 +10,9 @@
     "dompurify": "^2.5.7",
     "js-yaml": "^4.1.0",
     "marked": "^4.3.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "tgui-core": "^1.5.4",
     "tgui-dev-server": "workspace:*",
     "tgui-polyfill": "workspace:*"
   },
