@@ -4,21 +4,20 @@
  * @license MIT
  */
 
-import fs from 'fs';
-import { basename } from 'path';
-import { createLogger } from '../logging.js';
-import { require } from '../require.js';
-import { resolveGlob } from '../util.js';
+import fs from 'node:fs';
+import { basename } from 'node:path';
 
-const SourceMap = require('source-map');
-const { parse: parseStackTrace } = require('stacktrace-parser');
+import { SourceMapConsumer } from 'source-map';
+import { parse as parseStackTrace } from 'stacktrace-parser';
+
+import { createLogger } from '../logging.js';
+import { resolveGlob } from '../util.js';
 
 const logger = createLogger('retrace');
 
-const { SourceMapConsumer } = SourceMap;
 const sourceMaps = [];
 
-export const loadSourceMaps = async (bundleDir) => {
+export async function loadSourceMaps(bundleDir) {
   // Destroy and garbage collect consumers
   while (sourceMaps.length !== 0) {
     const { consumer } = sourceMaps.shift();
@@ -36,9 +35,9 @@ export const loadSourceMaps = async (bundleDir) => {
     }
   }
   logger.log(`loaded ${sourceMaps.length} source maps`);
-};
+}
 
-export const retrace = (stack) => {
+export function retrace(stack) {
   if (typeof stack !== 'string') {
     logger.log('ERROR: Stack is not a string!', stack);
     return stack;
@@ -81,4 +80,4 @@ export const retrace = (stack) => {
     })
     .join('\n');
   return header + '\n' + mappedStack;
-};
+}
