@@ -14,14 +14,16 @@ try {
 } catch {}
 
 // Timestamping function
-const getTimestamp = () => {
+function getTimestamp() {
   const timestamp = String(Date.now() - inception)
     .padStart(4, '0')
     .padStart(7, ' ');
   const seconds = timestamp.substr(0, timestamp.length - 3);
   const millis = timestamp.substr(-3);
+
   return `${seconds}.${millis}`;
-};
+
+}
 
 const getPrefix = (() => {
   if (isNode) {
@@ -41,22 +43,27 @@ const getPrefix = (() => {
     };
     return (ns) => [`%c${getTimestamp()}%c ${ns}`, styles.dimmed, styles.bright];
   }
+
   return (ns) => [`${getTimestamp()} ${ns}`];
 })();
 
 /**
  * Creates a logger object.
  */
-export const createLogger = (ns) => ({
-  log: (...args) => console.log(...getPrefix(ns), ...args),
-  trace: (...args) => console.trace(...getPrefix(ns), ...args),
-  debug: (...args) => console.debug(...getPrefix(ns), ...args),
-  info: (...args) => console.info(...getPrefix(ns), ...args),
-  warn: (...args) => console.warn(...getPrefix(ns), ...args),
-  error: (...args) => console.error(...getPrefix(ns), ...args),
-});
+export function createLogger(ns) {
+  return {
+    log: (...args) => console.log(...getPrefix(ns), ...args),
+    trace: (...args) => console.trace(...getPrefix(ns), ...args),
+    debug: (...args) => console.debug(...getPrefix(ns), ...args),
+    info: (...args) => console.info(...getPrefix(ns), ...args),
+    warn: (...args) => console.warn(...getPrefix(ns), ...args),
+    error: (...args) => console.error(...getPrefix(ns), ...args),
+  };
+}
 
 /**
  * Explicitly log with chosen namespace.
  */
-export const directLog = (ns, ...args) => console.log(...getPrefix(ns), ...args);
+export function directLog(ns, ...args) {
+  console.log(...getPrefix(ns), ...args);
+}
