@@ -4,9 +4,10 @@
  * @license MIT
  */
 
-import fs from 'fs';
-import os from 'os';
-import { basename } from 'path';
+import fs from 'node:fs';
+import os from 'node:os';
+import { basename } from 'node:path';
+
 import { DreamSeeker } from './dreamseeker.js';
 import { createLogger } from './logging.js';
 import { resolveGlob, resolvePath } from './util.js';
@@ -30,7 +31,7 @@ const SEARCH_LOCATIONS = [
 
 let cacheRoot;
 
-export const findCacheRoot = async () => {
+export async function findCacheRoot() {
   if (cacheRoot) {
     return cacheRoot;
   }
@@ -58,27 +59,30 @@ export const findCacheRoot = async () => {
     }
   }
   logger.log('found no cache directories');
-};
+}
 
-const onCacheRootFound = (cacheRoot) => {
+function onCacheRootFound(cacheRoot) {
   logger.log(`found cache at '${cacheRoot}'`);
   // Plant a dummy browser window file, we'll be using this to avoid world topic. For byond 514.
-  fs.closeSync(fs.openSync(cacheRoot + '/dummy', 'w'));
+  fs.closeSync(fs.openSync(cacheRoot + '/dummy.htm', 'w'));
 };
 
-export const reloadByondCache = async (bundleDir) => {
+export async function reloadByondCache(bundleDir) {
   const cacheRoot = await findCacheRoot();
   if (!cacheRoot) {
     return;
   }
   // Find tmp folders in cache
-  const cacheDirs = await resolveGlob(cacheRoot, './tmp*');
+  const cacheDirs = resolveGlob(cacheRoot, './tmp*');
   if (cacheDirs.length === 0) {
     logger.log('found no tmp folder in cache');
     return;
   }
-  // Get dreamseeker instances
-  const pids = cacheDirs.map((cacheDir) => parseInt(cacheDir.split('/cache/tmp').pop(), 10));
+
+  const pids = cacheDirs.map((cacheDir) => {
+    return parseInt(cacheDir.split('\\cache\\tmp')[1], 10);
+  });
+
   const dssPromise = DreamSeeker.getInstancesByPids(pids);
   // Copy assets
   const assets = await resolveGlob(bundleDir, './*.+(bundle|chunk|hot-update).*');
@@ -86,8 +90,8 @@ export const reloadByondCache = async (bundleDir) => {
     // Clear garbage
     const garbage = await resolveGlob(cacheDir, './*.+(bundle|chunk|hot-update).*');
     try {
-      // Plant a dummy browser window file, we'll be using this to avoid world topic. For byond 515.
-      fs.closeSync(fs.openSync(cacheDir + '/dummy', 'w'));
+      // Plant a dummy browser window file, we'll be using this to avoid world topic. For byond 515-516.
+      fs.closeSync(fs.openSync(cacheDir + '/dummy.htm', 'w'));
 
       for (let file of garbage) {
         fs.unlinkSync(file);
@@ -115,4 +119,4 @@ export const reloadByondCache = async (bundleDir) => {
       });
     }
   }
-};
+}
