@@ -4,9 +4,9 @@
  * @license MIT
  */
 
-import fs from 'fs';
-import { createRequire } from 'module';
-import { dirname } from 'path';
+import fs from 'node:fs';
+import { createRequire } from 'node:module';
+
 import { loadSourceMaps, setupLink } from './link/server.js';
 import { createLogger } from './logging.js';
 import { reloadByondCache } from './reloader.js';
@@ -18,17 +18,19 @@ const logger = createLogger('rspack');
  * @param {any} config
  * @return {RspackCompiler}
  */
-export const createCompiler = async (options) => {
+export async function createCompiler(options) {
   const compiler = new RspackCompiler();
   await compiler.setup(options);
+
   return compiler;
-};
+}
 
 class RspackCompiler {
   async setup(options) {
     // Create a require context that is relative to project root
     // and retrieve all necessary dependencies.
-    const requireFromRoot = createRequire(dirname(import.meta.url) + '/../..');
+    const requireFromRoot = createRequire(import.meta.dirname + '/../../..');
+    /** @type {typeof import('@rspack/core')} */
     const rspack = await requireFromRoot('@rspack/core');
     const createConfig = await requireFromRoot('./rspack.config.cjs');
     const config = createConfig({}, options);
@@ -75,7 +77,7 @@ class RspackCompiler {
         return;
       }
       stats
-        .toString(this.config.devServer.stats)
+        ?.toString(this.config.devServer.stats)
         .split('\n')
         .forEach((line) => logger.log(line));
     });
