@@ -7,4 +7,6 @@
 import { useSelector } from 'tgui/backend';
 import { selectDebug } from './selectors';
 
-export const useDebug = () => useSelector(selectDebug);
+export function useDebug() {
+  return useSelector(selectDebug);
+}
