@@ -28,27 +28,32 @@ const COLORS_SPECTRUM = [
 
 const COLORS_STATES = ['good', 'average', 'bad', 'black', 'white'];
 
-const Story = (props) => {
+function Story() {
   return (
     <Section>
       <Box mb={1}>
-        <Button content="Simple" />
-        <Button selected content="Selected" />
-        <Button altSelected content="Alt Selected" />
-        <Button disabled content="Disabled" />
-        <Button color="transparent" content="Transparent" />
-        <Button icon="cog" content="Icon" />
+        <Button>Simple</Button>
+        <Button selected>Selected</Button>
+        <Button disabled>Disabled</Button>
+        <Button color="transparent">Transparent</Button>
+        <Button icon="cog">Icon</Button>
         <Button icon="power-off" />
-        <Button fluid content="Fluid" />
-        <Button my={1} lineHeight={2} minWidth={15} textAlign="center" content="With Box props" />
+        <Button fluid>Fluid</Button>
+        <Button my={1} lineHeight={2} minWidth={15} textAlign="center">
+          With Box props
+        </Button>
       </Box>
       <Box mb={1}>
         {COLORS_STATES.map((color) => (
-          <Button key={color} color={color} content={color} />
+          <Button key={color} color={color}>
+            {color}
+          </Button>
         ))}
         <br />
         {COLORS_SPECTRUM.map((color) => (
-          <Button key={color} color={color} content={color} />
+          <Button key={color} color={color}>
+            {color}
+          </Button>
         ))}
         <br />
         {COLORS_SPECTRUM.map((color) => (
@@ -59,4 +64,4 @@ const Story = (props) => {
       </Box>
     </Section>
   );
-};
+}
