@@ -12,8 +12,9 @@ export const meta = {
   render: () => <Story />,
 };
 
-const Story = (props) => {
-  const [theme, setTheme] = useLocalState('kitchenSinkTheme');
+function Story() {
+  const [theme, setTheme] = useLocalState('kitchenSinkTheme', '');
+
   return (
     <Section>
       <LabeledList>
@@ -23,4 +24,4 @@ const Story = (props) => {
       </LabeledList>
     </Section>
   );
-};
+}
